backend/requirements.txt
Flask==3.0.3
python_dotenv==1.0.1